Refactored MitgliedDummy.java into MitgliedImpl.java
