+++ src/de/jost_net/JVerein/gui/view/MailVorlageDetailView.java
@@ -20,7 +20,7 @@ import de.jost_net.JVerein.Variable.MitgliedMap;
 import de.jost_net.JVerein.gui.action.DokumentationAction;
 import de.jost_net.JVerein.gui.action.OpenInsertVariableDialogAction;
 import de.jost_net.JVerein.gui.control.MailVorlageControl;
-import de.jost_net.JVerein.util.MitgliedDummy;
+import de.jost_net.JVerein.server.MitgliedImpl;
 import de.willuhn.jameica.gui.AbstractView;
 import de.willuhn.jameica.gui.GUI;
 import de.willuhn.jameica.gui.parts.ButtonArea;
@@ -44,7 +44,7 @@ public class MailVorlageDetailView extends AbstractView
     SimpleContainer t = new SimpleContainer(getParent(), true);
     t.addPart(control.getTxt());
 
-    Map<String, Object> map = new MitgliedMap().getMap(new MitgliedDummy(),
+    Map<String, Object> map = new MitgliedMap().getMap(MitgliedImpl.getDummy(),
         null);
 
     ButtonArea buttons = new ButtonArea();
